btrup: send/receive-based incremental backup utility
btrup uses the send and receive functionality to create incremental backups.
This replaces the rsync-based backup script.

Currently, the output from 'btrfs send' is written to a file on the backup
drive, then read by 'btrfs receive' to recreate the subvolume; this is because
piping the output directly seems to cause lockups.